Split datasets into datasets, streams and iterator
