+++ blocks/datasets/mnist.py
@@ -6,13 +6,13 @@ import numpy
 import theano
 
 from blocks import config
-from blocks.datasets import InMemoryDataset, lazy_properties
+from blocks.datasets import InMemoryDataset
 from blocks.datasets.schemes import SequentialScheme
 MNIST_IMAGE_MAGIC = 2051
 MNIST_LABEL_MAGIC = 2049
 
 
-@lazy_properties('features', 'targets')
+@InMemoryDataset.lazy_properties('features', 'targets')
 class MNIST(InMemoryDataset):
     u"""The MNIST dataset of handwritten digits.
