+++ docs/plotting.rst
@@ -40,7 +40,8 @@ function :math:`f(x) = x^a` to :math:`f(x) = x^2`.
 We train on a 150 random points in :math:`[0, 1]`.
 
 >>> import numpy
->>> from blocks.datasets import DataStream, ContainerDataset
+>>> from blocks.datasets.streams import DataStream
+>>> from blocks.datasets import ContainerDataset
 >>> sample = theano.tensor.scalar('data')
 >>> data_stream = DataStream(ContainerDataset(
 ...     numpy.random.rand(150).astype(theano.config.floatX)))
